Place shell vars in the right file
.zshrc is only sourced by interactive shells, where as .zshenv is
sourced by all shells.

The driving force of this is emacs' exec-file-path-from-shell
which ensures that the emacs environment matches that of the shell.
It does so by launching a non-interactive shell and copying the PATH so
we need to ensure that the PATH is set correctly even for
non-interactive shells.